+++ CHANGELOG.md
@@ -5,6 +5,9 @@ All notable changes to the Julia extension will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
+### Added
+- The `julia.environmentVariables` setting can now be used to configure the environment all Julia process start with ([#4031](https://github.com/julia-vscode/julia-vscode/pull/4031))
+- 
 ### Changed
 - The juliaup auto installer now uses `wget` by default (on Linux, with a fallback to `curl`) and tries the app installer and winget stores on Windows if the Microsoft store installation fails ([#4030](https://github.com/julia-vscode/julia-vscode/pull/4030))